Final v2.0

+++ backend/src/main/java/mkkz7/utils/CheckerInterface.java
@@ -1,5 +1,5 @@
 package mkkz7.utils;
 
 public interface CheckerInterface {
-    boolean calculate(int x, double y, int r);
+    boolean calculate(double x, double y, int r);
 }
